Use stable branch for deployment
DIT jenkins does not allow us to use webhooks to trigger builds as it is behind a vpn.Due to that reasosn Jenkins polling is used.

Prior to this change a tag named stable were used to mark stable version to deploy as live.Since jenkins polling can not detect changes on a tag it does not fire a build when tag is changed to point to another version. Hence, a branch for tracking the stable version is created for release/rollback purposes.